+++ src/actions/synchronization/toodledo/ToodledoGoalActions.js
@@ -23,7 +23,7 @@ export function synchronizeGoals() {
             const result = await Promise.all(goalsToAddPromises);
 
             for (let goal of result) {
-                await dispatch(updateGoal(goal, { loaded: true }));
+                await dispatch(updateGoal(goal, { loaded: true, skipUpdateMiddleware: true }));
 
                 if (goal.contributesTo) {
                     createdGoalsWithContributesTo.push(goal);
@@ -59,7 +59,7 @@ export function synchronizeGoals() {
                         await dispatch(addGoal(remoteGoal, { keepRefIds: true }));
                     } else {
                         if (!createdGoalsWithContributesTo.find(goal => goal.id === localGoal.id)) {
-                            await dispatch(updateGoal(merge(localGoal, remoteGoal), { loaded: true }));
+                            await dispatch(updateGoal(merge(localGoal, remoteGoal), { loaded: true, skipUpdateMiddleware: true }));
                         }
                     }
                 }
@@ -90,7 +90,7 @@ export function synchronizeGoals() {
             await Promise.all(goalsToUpdatePromises);
 
             for (let goal of goalsToUpdate) {
-                await dispatch(updateGoal(goal, { loaded: true }));
+                await dispatch(updateGoal(goal, { loaded: true, skipUpdateMiddleware: true }));
             }
         }
     };