+++ src/components/notefilters/ModalNoteFilterManager.js
@@ -9,6 +9,7 @@ function ModalNoteFilterManager() {
 
     const onCloseNoteFilterManager = () => {
         appApi.setNoteFilterManagerOptions({ visible: false });
+        appApi.refreshSelectedNoteFilter();
     };
 
     const onNoteFilterSelection = noteFilterId => {