+++ src/components/taskfilters/ModalTaskFilterManager.js
@@ -9,6 +9,7 @@ function ModalTaskFilterManager() {
 
     const onCloseTaskFilterManager = () => {
         appApi.setTaskFilterManagerOptions({ visible: false });
+        appApi.refreshSelectedTaskFilter();
     };
 
     const onTaskFilterSelection = taskFilterId => {